services/worker/requirements.txt
torch
torchvision
ultralytics
opencv-python
numpy
redis
pyyaml
python-dotenv
openai
transformers
pillow
